refactor: re-structure-folder
